+++ assets/js/notifications.js
@@ -3,6 +3,9 @@ import {awca_toast} from "./functions";
 jQuery(document).ready(function() {
     function fetchNotifications (page, limit){
         var anarNotifications =  jQuery('#awca_notification_list')
+        const markAsReadBtn = jQuery('#mark-page-as-read-btn')
+        const markAsReadBtnSpinner = markAsReadBtn.find('.spinner-loading');
+
         if(anarNotifications.length !== 0){
 
             var loadingIcon = anarNotifications.find('.spinner-loading')
@@ -13,19 +16,20 @@ jQuery(document).ready(function() {
                 type: "POST",
                 dataType: "json",
                 data: {
-                    action: 'awca_fetch_notifications_ajax',
+                    action: 'anar_fetch_notifications_ajax',
                     page: page,
                     limit: limit
                 },
                 beforeSend: function () {
                     loadingIcon.show();
+                    markAsReadBtn.prop('disabled', true);
                 },
                 success: function (response) {
                     if (response.success) {
                         anarNotifications.html(response.data.output)
                         msgType = 'success'
-                        mark_page_as_read(page, limit)
                         paginateNotifications(response.data.total, response.data.page, response.data.limit)
+                        updateMarkAsReadBtn(response.data.page, response.data.limit);
                     }
                     awca_toast(response.data.message, msgType);
                 },
@@ -36,8 +40,16 @@ jQuery(document).ready(function() {
                 },
                 complete: function () {
                     loadingIcon.hide();
+                    markAsReadBtn.prop('disabled', false);
                 },
             });
+
+            markAsReadBtn.on('click', function(e){
+                const page = jQuery(this).data('page');
+                const limit = jQuery(this).data('limit');
+                mark_page_as_read(page, limit)
+            })
+
 
             function paginateNotifications(total, page, limit) {
                 var pagination = jQuery('#awca_pagination');
@@ -56,6 +68,10 @@ jQuery(document).ready(function() {
                 });
             }
 
+            function updateMarkAsReadBtn(page, limit) {
+                markAsReadBtn.attr('data-page', page);
+                markAsReadBtn.attr('data-limit', limit);
+            }
 
             function mark_page_as_read(page, limit) {
                 var msgType = 'error'
@@ -64,18 +80,26 @@ jQuery(document).ready(function() {
                     type: "POST",
                     dataType: "json",
                     data: {
-                        action: 'awca_mark_as_read_notifications_ajax',
+                        action: 'anar_mark_as_read_notifications_ajax',
                         page: page,
                         limit: limit
                     },
+                    beforeSend: function () {
+                        markAsReadBtnSpinner.show()
+                        markAsReadBtn.prop('disabled', true);
+                    },
                     success: function (response) {
-                        if (response.success) {
+                        if(response.success) {
                             msgType = 'success'
                         }
                         awca_toast(response.data.message, msgType);
                     },
                     error: function (xhr, status, err) {
                         awca_toast(xhr.responseText)
+                    },
+                    complete: function () {
+                        markAsReadBtnSpinner.hide()
+                        markAsReadBtn.prop('disabled', false);
                     }
                 });
             }
@@ -85,4 +109,6 @@ jQuery(document).ready(function() {
     }
 
     fetchNotifications(1,100)
+
+
 })